Policies
- All policies from user, role, permissions up to research implemented

+++ app/Http/Controllers/Research/CitationController.php
@@ -25,6 +25,8 @@ class CitationController extends Controller
      */
     public function index(Research $research)
     {
+        $this->authorize('viewAny', ResearchCitation::class);
+
         $researchcitations = ResearchCitation::where('research_code', $research->research_code)->get();
 
         $research= Research::where('research_code', $research->research_code)->where('user_id', auth()->id())
@@ -40,6 +42,8 @@ class CitationController extends Controller
      */
     public function create(Research $research)
     {
+        $this->authorize('create', ResearchCitation::class);
+
         $researchFields = ResearchField::where('research_fields.research_form_id', 5)->where('is_active', 1)
             ->join('field_types', 'field_types.id', 'research_fields.field_type_id')
             ->select('research_fields.*', 'field_types.name as field_type_name')
@@ -55,6 +59,8 @@ class CitationController extends Controller
      */
     public function store(Request $request, Research $research)
     {
+        $this->authorize('create', ResearchCitation::class);
+
         $input = $request->except(['_token', '_method', 'document']);
 
         $id = ResearchCitation::insertGetId($input);
@@ -95,6 +101,8 @@ class CitationController extends Controller
      */
     public function show(Research $research, ResearchCitation $citation)
     {
+        $this->authorize('view', ResearchCitation::class);
+
         $researchFields = ResearchField::where('research_fields.research_form_id', 5)
                 ->join('field_types', 'field_types.id', 'research_fields.field_type_id')->where('is_active', 1)
                 ->select('research_fields.*', 'field_types.name as field_type_name')
@@ -119,6 +127,8 @@ class CitationController extends Controller
      */
     public function edit(Research $research, ResearchCitation $citation)
     {
+        $this->authorize('update', ResearchCitation::class);
+
         $researchFields = ResearchField::where('research_fields.research_form_id', 5)
                 ->join('field_types', 'field_types.id', 'research_fields.field_type_id')->where('is_active', 1)
                 ->select('research_fields.*', 'field_types.name as field_type_name')
@@ -146,6 +156,8 @@ class CitationController extends Controller
      */
     public function update(Request $request, Research $research, ResearchCitation $citation)
     {
+        $this->authorize('update', ResearchCitation::class);
+
         $input = $request->except(['_token', '_method', 'document']);
 
         $citation->update($input);
@@ -186,6 +198,8 @@ class CitationController extends Controller
      */
     public function destroy(Research $research, ResearchCitation $citation)
     {
+        $this->authorize('delete', ResearchCitation::class);
+
         $citation->delete();
 
         return redirect()->route('research.citation.index', $research->id)->with('success', 'Research Citation Deleted Successfully');
